dtypes: per-field goldens + mixed-policy E2E gates — feature complete
Increment B of docs/notes/dtype-policy-design.md (now marked
IMPLEMENTED). The goldens drop their flat-bf16-leaf trick for per-FIELD
typed leaves so non-default policies are golden-validated end to end:

- GoldenLlama3/Qwen3/Qwen35: parameters are {field: leaf} dicts at the
  policy's param dtypes (unpack_leaves); AdamW mirrors the runtime
  executor per field — gradient rounds through its grad STORAGE dtype,
  moments live at the opt dtype (_adamw_obj); tied qwen35 stays
  policy-addressed as head.*. Subclasses override layout hooks instead
  of view helpers; new final_leaves(object_id) is the gate-side
  comparison unit.
- gradcheck: ladder-2 uses per-field random state at storage dtypes and
  per-field dW at grad dtypes; check_model_step compares runtime W
  objects to golden leaves FIELD-BY-FIELD (dtype-true, padding-blind) —
  also retires the flat-bf16 comparison's fragility on gapped layouts.
- rmsnorm_reference outputs the ACTIVATION dtype even when the weight
  is stored wider (kernel semantics; bit-identical for bf16 weights) —
  fp32 norm weights no longer promote the residual stream.
- New acceptance gates (tests/training/test_dtype_policy_e2e.py):
  mixed-policy layouts, llama ladder-2 + model-step, and qwen35 tiny
  model-step with fp32 A_log/dt_bias/norms + fp32 moments everywhere —
  real engine vs golden, same tolerances as the bf16 ladders.

v1 scope notes in the design doc: fp32 on GEMM/table fields raises
loudly (cast-at-use unplumbed); ga>1 grad-accum rounding caveat
documented. Suite: 153 green.

+++ src/dataflow/tasks/ops.py
@@ -60,7 +60,9 @@ def rmsnorm_bwd(
 def rmsnorm_reference(x: torch.Tensor, w: torch.Tensor) -> torch.Tensor:
     xf = x.float()
     rstd = torch.rsqrt(xf.pow(2).mean(-1, keepdim=True) + RMS_EPS)
-    return (xf * rstd).to(x.dtype) * w
+    # output stays at the ACTIVATION dtype (kernel semantics) even when the
+    # weight is stored wider (dtype-policy fp32 norm weights)
+    return ((xf * rstd).to(x.dtype) * w).to(x.dtype)
 
 
 def rmsnorm_noweight(x: torch.Tensor) -> tuple[torch.Tensor, torch.Tensor]:
